The record id should be a long integer.

+++ src/main/scala/com/xebia/rest/Record.scala
@@ -2,7 +2,7 @@ package com.xebia.rest
 
 import cc.spray.json.DefaultJsonProtocol
 
-case class Record(id: Int,
+case class Record(id: Long,
                   shortStringAttribute: String,
                   longStringAttribute: String,
                   intNumber: Int,
